// web/src/gen/proto/orbit/v1/common/log_stream.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               unknown
// source: orbit/v1/common/log_stream.proto

/* eslint-disable */

export const protobufPackage = "orbit.v1.common";

export interface LogStreamEvent {
  type: string;
  source_id: string;
  cursor: string;
  data_base64: string;
  status: string;
  message: string;
  code: string;
  request_id: string;
  http_status: number;
  retryable: boolean;
}